bugfix: Don't download source/docs during Gradle extraction step
The step no longer takes 30 minutes on Windows unless you are in Debug mode (which is an advanced option). Although the 30 minutes only happens sometimes ... it seems to be related to whether Gradle is already running (which slows down the copying into the Gradle directory).

The actual fix is to remove the ALL keyword from dksdk.gradle.download. The ALL keyword means download documentation and source code ... which is not necessary.

fixes scoutapps#33